Create the server
First, imported the http module using require() built-in in node and
store it in a constant.

Then, use the http method createServer() to create a event listener
to gat request and manage response. Store it in a constant.

In that constant, put it to listen to a port by using the method
listen() and the server is started and in the event loop, a core aspect
of node