Create a route request
Inside the createServer, in the anonimous function, I created a if
statement to check the url. If the url is the root "/", the server would
send back a basic form with an input field to the browser. When the
sends a message, the form redirect to the "/message" page (through the
action parameter) and the if statement finds false and runs the other
block of code with the hello from the server.