getting access to the data written and puting it on the file
I created two new event listener. One for data and another for the end
The data comes to the server as stream and is available to be read
in the Buffer object. The event listener to streaming data pushes new
the upcoming data to the body array. Then, the req.on('end') event
listener uses the Buffer global object to concat the body array and
cast it as string. I consoled.log the string and write a .txt file
with that string in it.