rearrange the req.on('end') event listener to not block the code
I'm applying the event drive architecture to allow node to create
threads for each event and keep the event loop runing. It allows Node
to create threads of event that can run later and keep the server
open for new requests

I changed the fs.writeFileSync() to fs.writeFile() to write the file
async and provide a third argument, an arrow function to run when the
file is written. It allows the server to keep in the event loop and
accepting new requests, which is ideal