sending html files as response
I send html files as response to certain http requests. Using the
res.sendFile() I enter the url, as string, to that sendFile() method.
I can create the url using the global variable __dirname or, as a better
approach, create a utility folder with a path file exporting a function
that gets the root folder of the app and then joining to create the
url.

It's quite complex, but it generates better url and it can be read from
 lunix and windows