Output dynamic data using pug special syntax #{}
When setting a res.render(), you set the file to use as template and,
as second argument, you can pass an object with data to be used in the
template

inside the template, you can show the data received from the
res.render() method by using the syntax #{} that, inside the {}, you
put the data. Keep in mind that all data comes inside one object, so
use dot notation.

In this case, I used prods.title in which prods is the key value I pick
when passing data with the res.render() method and the title is the
inside the object received from the shop.js file.