add a id field in the save method inside the class
I created a this.id = Math.random().toString(); method inside the save() method in the model. That way, everytime
a new object is created, an id field is created and a random value is assigned. The Math.random() isn't
quite appropriate, but it's just dummy data. And then, I created an anchor tag that
points to a /product/id field being dynamically inserted by ejs.