add to cart functionality
I implement the add to cart functionality by creating a post request with data in the
request's body. Then, the data is accessed in the router using the req.body object.
There is another issue. When using include in the ejs template in a loop, you need
to pass the local variable as secong argument, inside an object, to make it available
in the loop.